+++ src/main/java/pigcart/cosycritters/loaders/fabric/ModMenuEntrypoint.java
@@ -3,13 +3,13 @@ package pigcart.cosycritters.loaders.fabric;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
-import pigcart.cosycritters.config.ConfigManager;
+import pigcart.cosycritters.config.gui.ConfigScreen;
 
 public class ModMenuEntrypoint implements ModMenuApi {
 
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return ConfigManager::screenPlease;
+        return ConfigScreen::screenPlease;
     }
 
 }